Update theme system
