+++ src/theme.ts
@@ -1,21 +1,14 @@
-import {createMuiTheme} from '@material-ui/core/styles';
 import {colors} from '../data/color';
 
-// Theme: https://material-ui.com/customization/theming/#theme-configuration-variables
-// Color: https://material-ui.com/customization/color/
-// Color-tool: https://material-ui.com/customization/color/#color-tool
-export const theme = createMuiTheme({
-  palette: {
-    primary: {
-      main: colors.DEPRECATED_primary,
-    },
-    secondary: {
-      main: colors.DEPRECATED_secondary,
-    },
+export const theme= {
+  color: {
+    primaryVivid: colors.primaryVivid,
+    white: colors.white,
+    baseLight: colors.baseLight,
+    baseDark: colors.baseDark,
   },
-  typography: {
-    button: {
-      textTransform: 'none',
-    },
+  layout: {},
+  fontFamily: {
+    primary: `-apple-system, BlinkMacSystemFont, Roboto, "Segoe UI", "Helvetica Neue", HelveticaNeue, YuGothic, "Yu Gothic Medium", "Yu Gothic", Verdana, Meiryo, sans-serif`
   },
-});
+} as const;
